Added stac search functions
